you canget the data from Kaggle at Link : https://www.kaggle.com/c/dogs-vs-cats-redux-kernels-edition/data
Data Description : 
The train folder contains 25,000 images of dogs and cats. Each image in this folder has the label as part of the filename. The test folder contains 12,500 images, named according to a numeric id. For each image in the test set, you should predict a probability that the image is a dog (1 = dog, 0 = cat).